==> R/00_all.qmd <==
---
title: "00_all"
format: html
editor: visual
message: false
---

```{r}
library(quarto)
```

```{r}
# Generation of the sub-documents when the main document is run
quarto::quarto_render("~/projects/Group_19_project/R/01_load.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/02_clean.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/03_augment.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/04_describe.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/05_analysis_mortality.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/06_analysis_comorbidities_etiologies.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/07_analysis_CPS.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/08_analysis_liver_disease.qmd")
quarto::quarto_render("~/projects/Group_19_project/R/09_logistic_analysis.qmd")
```

# Introduction

The aim of this project is to study the early predictors of cirrhosis' patients mortality. To do that, we investigate the effect of the different comorbidities, etiologies and liver diseases on the mortality, the Charlson Comorbidity Index and the Child Pugh Score of the patients.

## Loading the data

The dataset we used comes from the study "Early predictors of outcomes of hospitalization for cirrhosis and assessment of the impact of race and ethnicity at safety-net hospitals". The data can be downloaded by following the guideline that can be found in the 'README.md'.

```{r}
 file.rename(from = "~/projects/Group_19_project/R/01_load.html",
             to = "~/projects/Group_19_project/results/01_load.html" )
```

## Cleaning the data

```{r}
 file.rename(from = "~/projects/Group_19_project/R/02_clean.html",
             to = "~/projects/Group_19_project/results/02_clean.html" )
```

## Augmenting the data

```{r}
 file.rename(from = "~/projects/Group_19_project/R/03_augment.html",
             to = "~/projects/Group_19_project/results/03_augment.html" )
```

## Describing the data

```{r}
 file.rename(from = "~/projects/Group_19_project/R/04_describe.html",
             to = "~/projects/Group_19_project/results/04_describe.html" )
```

## Data analysis

```{r}
 file.rename(from = "~/projects/Group_19_project/R/05_analysis_mortality.html",
             to = "~/projects/Group_19_project/results/05_analysis_mortality.html" )
```

```{r}
 file.rename(from = "~/projects/Group_19_project/R/06_analysis_comorbidities_etiologies.html",
             to = "~/projects/Group_19_project/results/06_analysis_comorbidities_etiologies.html" )
```

```{r}
 file.rename(from = "~/projects/Group_19_project/R/08_analysis_liver_disease.html",
             to = "~/projects/Group_19_project/results/08_analysis_liver_disease.html" )
```

```{r}
 file.rename(from = "~/projects/Group_19_project/R/09_logistic_analysis.html",
             to = "~/projects/Group_19_project/results/09_logistic_analysis.html" )
```
