==> R/01_load.qmd <==
---
title: "01_load"
format: html
editor: visual
---

## Load library

```{r}
library("tidyverse")
```

## Load data

```{r}
dat_load_01 <- read_csv(file = 
        "~/projects/Group_19_project/data/_raw/PONE-D-18-29048R1_data_file.csv",
        na = c("_","NA", "NOT AVAILABLE", " ", ""))
```

```{r}
write_tsv(
  x= dat_load_01, 
  file = '~/projects/Group_19_project/data/01_dat_load.tsv'
)
```
