==> R/04_describe.qmd <==
---
title: "04_describe"
format: html
editor: visual
---

```{r}
library(tidyverse)
library(ggplot2)
library(patchwork)

```

```{r}
dat_load_01 <- read_tsv(file = 
        "~/projects/Group_19_project/data/_raw/PONE-D-18-29048R1_data_file.csv")
      
dat_aug_clean_03 <- read_tsv(file = 
        "~/projects/Group_19_project/data/03_dat_aug_clean.tsv") 

dat_aug_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug.tsv")

dat_aug_C_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug_C.tsv")
dat_aug_E_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug_E.tsv")
```

# Descriptive statistics

## General information on the data sets

We have 3 main data sets: the original one, the cleaned one with amongst other
things the removal of NA and the tidy data sets.

```{r}
# Original data set 
dim_desc(dat_load_01)
# Clean data set 
dim_desc(dat_aug_clean_03)
# Tidy data set 
dim_desc(dat_aug_03)
```

The original data set consisted in 733 cirrhosis' hospitalized patients with 91 
variables. After cleaning, and getting dropping non available observations, we
managed to reduce to 468 observations with 32 variables. Finally, we tidied 
using a pivot_longer to finally obtain a data set of 1065 observations over 9 
variables

## Sex and Age group of the patients' population

The total number of patients stratified by gender can be visualized with the 
following plots:

To better illustrate that both have their mode (most commonly occurring number 
in the data set) at \[60;70\], the plot is also shown faceted.

```{r}
bar_plot_age_gender <- dat_aug_clean_03 |>
  ggplot(aes(x = Age, 
             fill = Sex)) +
  geom_bar(position = "stack", 
           stat = "count", 
           alpha = 0.7) +
  labs(title = "     Overview of the age and gender distribution of patients with 
       liver cirrhosis",
       x = "Age Group",
       y = "Count") +
  scale_fill_viridis_d() +
#  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5, face = "bold", size = 14)) +
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=14,face="bold"))

bar_plot_age_gender
```

### Save plot

```{r}
ggsave("Age_gender.png", 
       path = "../results/plots/",
       plot = bar_plot_age_gender,
       width = 9,
       height = 5)
```

The plots provide a visual representation of the distribution of patients with
cirrhosis based on age and gender.

-   the x-axis always represents the age groups and the y-axis is always the 
count for the patients

-   the gender is illustrated by different colors - yellow for male and purple
for female

```{r}
mode_of_age <- dat_aug_clean_03 |> 
  filter(str_detect(Age, "\\[60;|70]" )) |> 
  summarise(count = n())
print(mode_of_age)
```

```{r}
mode_of_age_sex <- dat_aug_clean_03 |> 
  filter(str_detect(Age, "\\[60;|70]" )) |> 
  group_by(Sex) |> 
  summarise(count = n())

mode_of_age_sex
```

```{r}
dat_aug_clean_03 |> 
  count(Sex)
```

-   **Mode**: The mode is for both women and men in the 60 to 70 age group. 
The number of patients in this age group is 217 (Female: 69, Male 148).

-   **Gender Representation**: Women make up about 32.69 % of patients and men 
about 67.31 %. This means that the number of patients is predominantly male.

## Overview of the liver disease in the patients' population

```{r}
dat_LD <- dat_aug_clean_03 |> 
  count(Liver_disease) |> 
  mutate(proportion = n/sum(n))
```

Ascites and hepatic encephalopathy are the two main liver diseases diagnosed.
Ascites, representing 31% of liver diseases, is a build up of fluid in the 
abdomen, often as a consequence of cirrhosis. Hepatic encephalopathy is also a 
common pathology in patients with cirrhosis due to deficiencies in their
neurological system, here representing 21.2% of the study.

### Bar plot version

```{r}
bar_plot_LD<- dat_LD |> 
  ggplot(aes(x = reorder(Liver_disease,-n), 
             y = n, 
             fill = Liver_disease)) +
  geom_bar(stat = 'identity', 
           alpha = 0.7, 
           show.legend = FALSE) +
  geom_text(aes(label = n), 
            vjust = -0.5,  
            color = "black", 
            size = 3) +  
  labs(title = "Liver disease repartition in cyrrhosis patients",
       y = "Count") +
  scale_fill_viridis_d() +
  theme(axis.title.x=element_blank()) +
  scale_x_discrete(guide = guide_axis(angle = 25))
```

### Pie chart version

```{r}
pie_chart_LD<- dat_LD |> 
  ggplot(aes(x = "", 
             y = proportion, 
             fill = Liver_disease)) +
  geom_bar(stat = 'identity', 
           alpha = 0.7) +
  coord_polar("y", start=0) +
  labs(title = "            Liver disease repartition in cirrhosis patients",
       y = "Count") +
  geom_text(aes(label = sprintf("%.1f%%", proportion * 100)),
            position = position_stack(vjust = 0.3))+
  scale_fill_viridis_d() +
  theme(axis.title.x=element_blank(), 
        axis.text.x = element_blank(), 
        axis.line.x = element_blank(), 
        axis.title.y=element_blank(), 
        axis.text.y = element_blank(), 
        axis.line.y = element_blank(), 
        axis.line = element_blank())+
  theme(plot.title = element_text(face = "bold", size = 14))
```

```{r}
pie_chart_LD
```

```{r}
ggsave("Overview_LD.png", 
       path = "../results/plots/",
       plot = pie_chart_LD,
       width = 9,
       height = 5)
```

## Overview of comorbities in the patients' population

Comorbidities are additional disease that can interact and coexist 
simultaneously with the cirrhosis. The study of comorbidities is pertinent since
it can explain some of the patients' complications and can play a role in their 
recovery.

```{r}
dat_comorb <- dat_aug_C_03 |> 
  group_by(Comorbidities) |> 
  summarise(count = n()) 

dat_comorb
```

```{r}
bar_plot_comorbidities <- dat_comorb |> 
  ggplot(aes(x = reorder(Comorbidities,-count), 
             y = count/sum(count),
             fill = Comorbidities)) +
  geom_bar(stat = 'identity', 
           alpha = 0.7, 
           show.legend = FALSE) +
  geom_text(aes(label = sprintf("%.1f%%", count/sum(count)*100)), 
            vjust = -0.5, 
            size = 2) +
  labs(title = "Overview of comorbidities across cirrhosis patients",
       y = "Proportion") +
  scale_fill_viridis_d() +
  theme(legend.title=element_blank(), 
        axis.title.x = element_blank(), 
        plot.title = element_text(hjust = 0.2, face = "bold", size = 14)) +
  scale_x_discrete(guide = guide_axis(angle = 63))

bar_plot_comorbidities
```

```{r}
ggsave("Overview_Comorbidities.png", 
       path = "../results/plots/",
       plot = bar_plot_comorbidities,
       width = 7,
       height = 5)
```

The plot illustrates the various concomitant diseases of the patients and 
provides an overview of the number of each condition.

-   the x-axis represents the comorbidities and the y-axis is the count of 
patients

-   In addition, each comorbidity is listed in different colors and the exact
proportion of the etiology in the population is indicated above the respective
bar.

-   The mode are moderately severe liver diseases with a comparatively high 
count of 46.8%

-   The majority of patients suffer from concomitant diseases. Only 8 patients 
suffer only from liver cirrhosis.

## Overview of ethiology factor across population

Ethiology is the study of the factors that come together leading to a disease.

```{r}
dat_etio <- dat_aug_E_03 |> 
  group_by(Etiologies) |> 
  summarise(count = n()) 

dat_etio
```

```{r}
bar_plot_etiology <- dat_etio |> 
  ggplot(aes(x = reorder(Etiologies,-count), 
             y = count/sum(count),
             fill = Etiologies)) +
  geom_bar(stat = 'identity', 
           alpha = 0.7, 
           show.legend = FALSE) +
  geom_text(aes(label = sprintf("%.1f%%", count/sum(count)*100)), 
            vjust = -0.5, 
            size = 3) +
  labs(title = "Overview of etiology across cirrhosis patients",
       y = "Proportion") +
  scale_fill_viridis_d() +
  theme(legend.title=element_blank(), 
        axis.title.x = element_blank(), 
        plot.title = element_text(hjust = 0.2, face = "bold", size = 14)) +
  scale_x_discrete(guide = guide_axis(angle = 30))

bar_plot_etiology
```

```{r}
ggsave("Overview_Etiology.png", 
       path = "../results/plots/",
       plot = bar_plot_etiology,
       width = 7,
       height = 5)
```

The plot illustrates the various etiologies which lead to diseases and counts
them.

-   the x-axis represents the etiologies and the y-axis is the count of patients

-   The etiologies are listed in different colors and their propotion in the 
population is indicated above the respective bar.

-   The mode is the consumption of alcohol with a proportion of 55 % followed
by hepatitis C at 33.2%. Both are likely major factors at causing liver 
cirrhosis.

## Overview of two important numerical values: CPS and Charlson_CI

The data set chosen featured two variables:

-   **CPS or Child-Pugh Score** for Cirrhosis Mortality: useful in the 
determination of the prognosis of patients suffering from a cirrhosis. With a 
maximum value of 15, it is based on the presence of ascites, the concentratin of
some proteins in the blood (bilirubin, albumin)...

-   **Charlson Comorbidity index**: useful to predict the mortality of patients 
suffering from comorbidities. Depends on the presence of comorbidities and the
age of the patient

```{r}
box_plot_CPS <- dat_aug_C_03 |> 
  ggplot(aes(x = CPS)) +
  geom_boxplot(alpha = 0.7) +
  labs(title = "CPS and Charlson_CI") +
  theme_minimal() +
  coord_flip()

box_plot_Charlson <- dat_aug_03 |> 
  ggplot(aes(x = Charlson_CI)) +
  geom_boxplot(alpha = 0.7) +
  theme_minimal() +
  coord_flip()

box_plot_CPS_Charlson <- box_plot_CPS | box_plot_Charlson 

box_plot_CPS_Charlson
```

```{r}
mean_CPS <- dat_aug_C_03 |> 
  summarize(mean(CPS))

mean_CPS
```

```{r}
mean_Charlson <- dat_aug_C_03 |> 
  summarize(mean(Charlson_CI))

mean_Charlson
```

The average value of the CPS and the Charlson comorbidity index are reciprocally 
8.6 and 3.83. It is in the group B of CPS which implies a 80% 1-year survival
rate and 60% 2-year survival rate.

### Relation between Charlson value and the type of comorbidities

Charlson value represents a comorbidity index that predicts 10y life expectancy 
for people with multiple comorbidities. This graph shows the weight of varioous 
Comorbidiities in the Charlson Index

```{r}

box_plot_Charlson_Comorbidities <- dat_aug_C_03 |> 
  ggplot(mapping = aes(x = Comorbidities,
                       y = Charlson_CI)) +
  geom_boxplot() +
  labs(title = 'Impact of diverse comorbidities on the 10 year life expectancy
       of patients',
       y = 'Charlson Comorbidity Index') +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))

box_plot_Charlson_Comorbidities
```

From the boxplot, we can deduce that AIDS and Metastatic solid tumor are 2 of 
the major comorbidities that have an important weight on the Charlson CI. On the
opposite side of the spectrum, mild liver disease and rheumatological have a 
limited impact on the variable. Less weight means that the diseaseis not severe 
enough to affect directly and significantly the life expectancy of the patient.

## Overview of comorbidities and liver disease

```{r}
stacked_bar_liverDisease <- dat_aug_C_03 |>
  select(Liver_disease, Comorbidities) |> 
  count(Liver_disease, Comorbidities)|>
  rename(d_c_count = n) |> 
  arrange(Liver_disease, desc(d_c_count)) |>
  mutate(Comorbidities = fct_inorder(Comorbidities))
```

```{r}
bubble_order <- c("UGIB due to portal hypertension",
                  "Jaundice- as isolated presentation",
                  "Hepatic encephalopathy",
                  "Hepato-hydrothorax",
                  "HCC or cholangioca.",
                  "Alcoholic hepatitis",
                  "Other fluid overload",
                  "Ascites",
                  "SBP",
                  "AKI")
bubble_chart <- stacked_bar_liverDisease |>
  arrange(match(Liver_disease,bubble_order))
```

```{r}
bubble_comorbiditie_by_LD <- ggplot(bubble_chart,
       aes(x = factor(Liver_disease, bubble_order), 
           y = Comorbidities,
           colour = Liver_disease,
           size = d_c_count)) +
  geom_point() +
  geom_text(aes(label = d_c_count), 
            colour = "white", 
            size = 3) +
  scale_x_discrete(position = "top") +
  scale_color_viridis_d() +
  scale_size_continuous(range = c(3, 9)) +
  labs(x = NULL, y = NULL, title= "Comorbidities by Liver Disease") +
  theme(legend.position = "none",
        panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.ticks = element_blank(),
        axis.text.x = element_text(angle = 45, hjust = 0))

bubble_comorbiditie_by_LD
```

From this visualisation graph, it appears that Hepatic Encephalopaty liver 
disease has the greatest amount of comorbidities observed. On the opposite, we 
observe fewer Comorbidities for for jaundice and Hepato-hydrothorax.
Additionally, the most common combination is Ascites with Moderate Liver disease
as a comorbidity.
