==> R/05_analysis_mortality.qmd <==
---
title: "05_analysis_mortality"
format: html
editor: visual
---

```{r}
library("tidyverse")
library("ggpubr")
library("ggplot2")
library("dplyr")
```

```{r}
dat_aug_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug.tsv")
dat_aug_C_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug_C.tsv")
dat_aug_E_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug_E.tsv")
```

# Effect of different variables on mortality

During that part, we will investigate to determine to what extent some variables like CPS, age and liver disease affect the mortality of the patients before investigating specific and relevant etiologies that have been picked using a linear regression performed in file 07_analysis_CPS.qmd.

## CPS Mortality

```{r}
bar_plot_CPS_Mortality<- dat_aug_03 |> 
  ggplot(aes(x = Mortality,
             y = CPS,
             fill = Mortality)) +
  geom_boxplot(alpha = 0.7) +
  labs(title = "Boxplot chart of CPS values",
       subtitle = "stratified by mortality",
       y = "CPS") +
  scale_fill_viridis_d() +
  theme_minimal() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank()) +
  scale_x_discrete(guide = guide_axis(angle = 25))

bar_plot_CPS_Mortality

```

This plot proves the fact that CPS value accurately describes the mortality of cirrhosis patients. It shows that a low CPS is correlated to a higher survival rate. On the other side, a patient with a high CPS has more chances of dying of cirrhosis.

### Saving the plot

```{r}
ggsave("CPS_and_mortality_boxplot.png", 
       path = "../results/plots/",
       plot = bar_plot_CPS_Mortality,
       width = 7,
       height = 5)
```

## Age impact on mortality

First, we will see how the age of the patients impact their mortality by plotting the distribution of patient outcomes stratified by age.

```{r}
Age_impact_mortality1 <- dat_aug_03 |> 
  ggplot(mapping = aes(x = Age,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "stack") +
  scale_fill_viridis_d() +
  labs(title = "Distribution of patient outcomes ",
       subtitle = "stratified by age",
       x = "Age",
       y = "Count") +
  theme_minimal()

Age_impact_mortality1
```

```{r}
Age_impact_mortality2 <- dat_aug_03 |> 
  ggplot(mapping = aes(x = Age,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "fill") +
  scale_fill_viridis_d() +
  labs(title = "Proportions of patient outcomes ",
       subtitle = "stratified by age",
       x = "Age",
       y = "Proportions"
)

Age_impact_mortality2
```

The plot showing the proportions of patient outcomes is more relevant than the one showing their distribution, because it is relative to the count of patients. Nonetheless, the distribution plot allows us to see that the age categories \[30,40\] and \[90,100\] don't contain enough patients to be statistically relevant. Therefore, if we only take the other age categories into consideration, we can see slight increase in mortality when the age of the patients is increasing.

### Saving the plots

```{r}
ggsave("Age_Mortality_count.png", 

       path = "../results/plots/",
       plot = Age_impact_mortality1,
       width = 7,
       height = 5)

ggsave("Age_Mortality_prop.png", 
       path = "../results/plots/",
       plot = Age_impact_mortality2,
       width = 7,
       height = 5)

```

## Effect of different liver diseases on mortality

```{r}
bar_plot_LD_Mortality<- dat_aug_03 |> 
  ggplot(aes(x = Liver_disease, 
             fill = Mortality)) +
  geom_bar(position = "fill", 
           stat = "count", 
           alpha = 0.7) +
  labs(title = "                          Liver disease in cirrhosis patients, stratified by mortality",
       y = "Proportion") +
  scale_fill_viridis_d() +
  theme(axis.title.x=element_blank()) +
  scale_x_discrete(guide = guide_axis(angle = 35)) +
  theme(plot.title = element_text(hjust = 0.5, face = "bold", size = 14)) +
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=14,face="bold"))

bar_plot_LD_Mortality
```

By looking at the mortality of patients with different liver diseases, we can see the type of liver disease impacts greatly the mortality of patients. The size of the samples "Hepato-hydrothorax" and "Jaundice- as isolated presentation" being small (less than 15 patients), the data concerning these liver disease should not be taken into account for the analysis. This plot also provides some information about how fulminant the different liver diseases are. For instance, the categories "Other fluid overload" and "SBP" have approximately the same proportions of overall mortality, but the category "Other fluid overload" seems to be more fulminant because all of the patients that died after contracting that disease died in hospital, when the majority of the patient that died after contracting SBP died more than 30 days after leaving the hospital.

### Saving the plot

```{r}
ggsave("LD_Mortality.png", 
       path = "../results/plots/",
       plot = bar_plot_LD_Mortality,
       width = 7,
       height = 5)
```

# Effect of different etiologies on mortality

We start by creating a new data set where we add three variables. The three variable are three different etiologies and their significance concerning Child Pugh score was assessed with a linear regression model in the file 07. Because the Child Pugh Score indicates the severity of the cirrhosis, it could be related to the mortality, and we will verify that hypothesis with the following plots.

```{r}
dat_mortality <- dat_aug_03 |> 
  mutate(Alcohol = ifelse(Etiologies == "Alcohol_consumption", "yes", "no")) |> 
  mutate(NAFLD = ifelse(Etiologies == "Non_alcoolic_fatty_LD", "yes", "no")) |> 
  mutate(Cryptogenic = ifelse(Etiologies == "Cryptogenic", "yes", "no")) 

dat_mortality

```

## Alcohol consumption impact on mortality

To determine if alcohol consumption impacts mortality, we plot the distribution and the proportions of patient outcomes stratified by alcohol consumption. Alcohol was considered by the linear regression as a non-significant etiology, we therefore expect to see very few changes in the proportions. We will here plot the distribution (with normal counts) and the proportion of the etiology in the population to be able to have a relative comparison.

```{r}
Alcohol_distribution <- dat_mortality |> 
  ggplot(mapping = aes(x = Alcohol,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "stack") +
  scale_fill_viridis_d() +
  labs(title = "Patient outcomes",
       subtitle = "stratified by alcohol consumption",
       x = "Alcohol consumption", 
       y = "Count") + 
  theme(legend.position = "none")


Alcohol_proportions <- dat_mortality |> 
  ggplot(mapping = aes(x = Alcohol,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "fill") +
  scale_fill_viridis_d() +
  labs(
    title = "Proportions of patient outcomes ",
    subtitle = "stratified by alcohol consumption",
    x = "Alcohol consumption",
    y = "Proportions")

Alcohol_impact_mortality <- ggarrange(Alcohol_distribution,
                                      Alcohol_proportions, 
                                      legend = 'bottom',
                                      common.legend = TRUE)
                                
Alcohol_impact_mortality 
```

The proportions plot shows a very slight/ unsignificant increase in mortality when the patients are alcohol consumers, therefore we can't notice a big mortality difference between the two populations.

### Saving the plot

```{r}
ggsave("Alcohol_impact_mortality.png", 
       path = "../results/plots/",
       plot = Alcohol_impact_mortality,
       width = 7,
       height = 5)
```

## Non-alcoholic fatty liver disease impact on mortality

On the contrary to alcohol consumption, NAFLD was considered a significant etiology.

```{r}
NAFLD_distribution <- dat_mortality |> 
  ggplot(mapping = aes(x = NAFLD,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "stack") +
  scale_fill_viridis_d() +
  labs(title = "Patient outcomes",
       subtitle = "stratified by presence of NAFLD",
       x = "Patient suffering from NAFLD", 
       y = "Count") + 
  theme(legend.position = "none")

NAFLD_proportions <- dat_mortality |> 
  ggplot(mapping = aes(x = NAFLD,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "fill") +
  scale_fill_viridis_d() +
  labs(title = "Proportions of patient outcomes",
       subtitle = "stratified by presence of NAFLD",
       x = "Patient suffering from NAFLD",
       y = "Proportions"
)

NAFLD_impact_mortality <- ggarrange(NAFLD_distribution,
                                      NAFLD_proportions, 
                                      legend = 'bottom',
                                      common.legend = TRUE)
NAFLD_impact_mortality
```

On the proportions plot, we can see that patients with NAFLD seem to have more chance to survive to cirrhosis. However, we can't reach any specific conclusion due to the small population of patients with NAFLD. NAFLD, significant etiology, has therefore a 'positive' impact on cirrhosis patients, which seems hard to understand.

### Saving the plot

```{r}
ggsave("NAFLD_impact_mortality.png", 
       path = "../results/plots/",
       plot = NAFLD_impact_mortality,
       width = 7,
       height = 5)
```

# Effect of the number of etiologies and comorbidities

Another hypothesis that can be made could be that the number of etiologies and comorbidities that a patient has impacts the mortality of that patient.

## Analysis number of etiologies

```{r}
dat_aug_E_03 <- dat_aug_E_03 |> 
    group_by(Id) |> 
    mutate(number_etiologies = n()) |> 
    ungroup()

dat_aug_E_03

```

```{r}
nb_etiologies <- dat_aug_E_03 |> 
  ggplot(mapping = aes(x = number_etiologies,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "fill") +
  scale_fill_viridis_d() +
  labs(title = "Patient outcomes",
       subtitle = "stratified by the number of etiologies",
       x = "Number of etiologies", 
       y = "Count") 
nb_etiologies

```

```{r}
ggsave("Number_of_E_mortality.png", 
       path = "../results/plots/",
       plot = nb_etiologies,
       width = 7,
       height = 5)
```

## Analysis number of comorbidities

```{r}
dat_aug_C_03 <- dat_aug_C_03 |> 
    group_by(Id) |> 
    mutate(number_comorbidities = n()) |> 
    ungroup()

dat_aug_C_03

```

```{r}
nb_comorbidities <- dat_aug_C_03 |> 
  ggplot(mapping = aes(x = number_comorbidities,
                       fill = Mortality,
                       na.rm = TRUE)) +
  geom_bar(alpha = 0.6,
           position = "fill") +
  scale_fill_viridis_d() +
  labs(title = "Patient outcomes",
       subtitle = "stratified by the number of comorbidities",
       x = "Number of comorbidities", 
       y = "Count") 
nb_comorbidities

```

```{r}
ggsave("Number_of_C_mortality.png", 
       path = "../results/plots/",
       plot = nb_comorbidities,
       width = 7,
       height = 5)
```

```{r}
distinct_E<- dat_aug_E_03 |>
  # to count the patients with many etiologies only once
  distinct(Id, .keep_all = TRUE) 

E_table <- distinct_E %>%
  group_by(number_etiologies) %>%
  summarise(number_of_patients = n())


E_table
```

```{r}
distinct_C<- dat_aug_C_03 |>
  # to count the patients with many comorbidities only once
  distinct(Id, .keep_all = TRUE) 

C_table <- distinct_C %>%
  group_by(number_comorbidities) %>%
  summarise(number_of_patients = n())


C_table
```

By looking at the number of patients present in each sample, we can see that there is only one patient that has six comorbidities, so we choose not to take that category into account for the analysis of the bar charts. The bar chart stratified on the number of etiologies shows that this number is proportional to the mortality: the more etiologies a patient has, the more likely he is to die of cirrhosis. Concerning the bar chart stratified on the number of comorbidities, there is no such trend in the mortality proportions, but we can still see that the proportion of death that occur in the hospital increases, so the severity of the disease might increase with the number of comorbidities.

## Conclusion

This analysis allows to conclude that CPS and the type of liver disease have a major impact on patient's outcome, whereas age has a slight impact. We also got to test the linear regression significant results, noticing that NAFLD seems to have a positive impact on the mortality, whereas alcohol consumption doesn't seem to play a role in the patients outcome. Regardless of the type of early predictors, their number also has an effect on the observed mortality. Indeed, an important number of etiologies tends to increase the mortality of the patients.
