==> R/06_analysis_comorbidities_etiologies.qmd <==
---
title: "06_analysis_comorbidities_etiologies"
format: html
editor: visual
---

```{r}
library("tidyverse")
library("ggpubr")
library("viridis")
library("ggplot2")
library("dplyr")
```

```{r}
dat_aug_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug.tsv")
```

# Liver disease and comordities/etiologies relations

## Liver disease and Comorbidities

Seeking to understand the relation between comorbidities and the patient's liver disease, it would be helpful to visualize if some liver diseases have peculiar relationships with some comorbidities. To do so, a tile plot with the variables Liver disease, Comorbidities and ratio of patients of a certain liver disease with comorbidities was created.

### Creation of a dataset with count of patients with same disease and same comorbidities

```{r}
grouped_LD_C <- dat_aug_03 |>
  group_by(Liver_disease, Comorbidities) |> 
  summarise(count = n()) |> 
  arrange(Liver_disease, desc(count)) 

grouped_LD_C
```

### Creation of a dataset with total count of patients with same disease

```{r}
grouped_LD <- dat_aug_03 |>
  group_by(Liver_disease) |> 
  summarise(count_LD = n())

grouped_LD
```

### Joining of the two dataset according to Liver_disease

```{r}
merged_dat1 <- inner_join(grouped_LD_C, grouped_LD, by = 'Liver_disease')

# Calculate the ratio of count to total count
merged_dat1 <- merged_dat1|>
  mutate(Ratio = count / count_LD)

merged_dat1
```

```{r}
tile_plot_comorb <- merged_dat1 |>
  ggplot(aes(Liver_disease, 
             Comorbidities)) +
  geom_tile(aes(fill= Ratio))+
  labs(fill = "Ratio" ,
       title = "Proportions of comorbidities for each liver disease ",
       x = "Liver Disease",
       y = "Comorbidities") +
  scale_fill_viridis() +
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), 
        plot.title = element_text(hjust = 0.73, face = "bold", size = 14))

tile_plot_comorb
```

```{r}
ggsave("Tile_plot_comorbidities.png", 
       path = "../results/plots/",
       plot = tile_plot_comorb,
       width = 7,
       height = 5)
```

Each column adds up to 100%. From this graph, we can state that Moderate to severe liver disease is an important comorbidities for all liver diseases. This comorbidity. meaning was left very vague in the study, so it is hard to give an actual explanation. However, what we can say is that close to 60% of patients suffering from a Hepato-hydrothorax have a moderate to severe liver diseasse in addition. A lot of Comorbidities are not clustered for a particular liver disease, except for Malignancy and Diabetes which seem to be present in around 30% of patients suffering respectively from HCC and Jaundice. On the other side of the spectrum Heiplegia/Paraplegia and Dementia have very low incidence in the cirrhosis' patients population.

# Liver disease and etiology

After tackling the subject comorbidities that can enhance some cases of cirrhosis, the etiology or causes of the liver disease is also a useful tool, both for treatment and prevention.

### Creation of a new dataset counting patients with same disease and same etiology

```{r}
grouped_LD_E <- dat_aug_03 |> 
  group_by(Liver_disease, Etiologies) |> 
  summarise(count = n()) 

grouped_LD_E
```

### Joining of the two dataset according to Liver_disease

```{r}
merged_dat2 <- inner_join(grouped_LD_E, grouped_LD, by = 'Liver_disease')

# Calculate the ratio of count to total count
merged_dat2 <- merged_dat2|>
  mutate(Ratio = count / count_LD)

merged_dat2
```

```{r}
tile_plot_etiologies <- ggplot(merged_dat2, 
       aes(x = Liver_disease, 
           y = Etiologies, 
           fill = Ratio)) +
  geom_tile()+
  labs(title = "Proportions of etiologies for each liver disease",
       x = "Liver Disease",
       y = "Etiologies", 
       fill = "Percent of patients") + 
  scale_fill_viridis()+
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), 
        plot.title = element_text(hjust = 0.50, face = "bold", size = 13))

tile_plot_etiologies

```

Each column adds up to 100%.

From this tile chart, Hepathitis C and alcohol consumption appear like the two main ethiologies for the liver disease, especially for Alcoholic hepatitis where nearly 80% is caused by alcohol consumption. Hepathitis C is also responsible for half of the cases of HCC of patients in this study. Those two ethiologies usually share the stage for each liver disease. Ascites, Hepatic encephalopathy and UGIB have multiple etiologies possible at few percentages like Autoimmune hepatitis, cryptogenic disease, Hepathitis B and Non Alcoholic fatty liver disease. However, Primary biliary cholangitis seems like a very rare etiology, since it is only present in a very few percent (arouond 1%) for patients suffering from Fluid overload.

# Does etiology have influences on the mortality of patients for the same liver disease ?

```{r}
dat_aug_03_2 <- dat_aug_03 |> 
  filter(Mortality != 'No death') |> 
  group_by(Liver_disease, Etiologies, Mortality) |> 
  summarise(count2 = n())
dat_aug_03_2
```

```{r}
dat_aug_03_join <- left_join(grouped_LD_E, 
                             dat_aug_03_2, 
                             by = c("Liver_disease",
                                    "Etiologies"))
```

```{r}
dat_aug_03_join <- dat_aug_03_join |> 
  replace_na(list(count2 = 0))
```

```{r}
tile_plot_mortality <- ggplot(dat_aug_03_join, 
       aes(x = Liver_disease, 
           y = Etiologies, 
           fill = count2/count)) +
  geom_tile() +
  labs(title = "Death ratios of Etiology/Liver disease combination ",
       x = "Liver Disease",
       y = "Etiologies", 
       fill = "Death ratio") +
  scale_fill_viridis()+
  theme_classic()+ 
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), 
        plot.title = element_text(hjust = 0.60, face = "bold", size = 13),
        axis.text=element_text(size=12),
        axis.title=element_text(size=14,face="bold"))

tile_plot_mortality
```

```{r}
ggsave("Tile_plot_mortality.png", 
       path = "../results/plots/",
       plot = tile_plot_mortality,
       width = 7,
       height = 5)
```

After inspecting the etiologies associations with the liver diseases, we used the tile chart above to show the death ratio for each liver disease and its associated etiology. From this plot, we can deduce that all patients that suffered from Hepatic encephalopathy with auto-immune hepatitis died of cyrrhosis. The combination AKI/cryptogenic disease (70%), Hepato-hydrothorax/Alcohol-consumption (80%), Jaundice/Hepathitis C (80%) have the strongest deatah ratio. On the opposite, it seems that ethiologies like Hepathitis B, Non-alcoolic fatty liver disease and liver disease like Alcoholic hepatitis, Ascites, HCC, Fluid overload, SBP and UGIB have lower dath ratio (less thann 25%).

# Conclusion

After plotting the proportions of comorbidities and etiologies for each liver disease, we noticed that some of these are over represented, for instance the comorbidity "Moderate to severe liver disease", as well as the etiologies "Hepathitis C" and "Alcohol consumption". Moreover, by plotting a tile chart of the liver diseases and the etiologies that takes the death ratio into account, we saw that some combinations of liver diseases and etiologies are particularly lethal, in particular the combination of autoimmune-hepatitis and hepatic-encephalopathy.
