==> R/07_analysis_CPS.qmd <==
---
title: "07_analysis_CPS"
format: html
editor: visual
---

```{r}
library("tidyverse")
library("ggpubr")
library("forcats")
library("purrr")
library("broom")
library("ggplot2")
library("dplyr")
```

```{r}
dat_aug_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug.tsv")
df_Etiology <- read_tsv("~/projects/Group_19_project/data/02_df_E.tsv")

```

# Effects of comorbidities on Child Pugh Score

The Child Pugh Score is a variable used to indicate the severity of a cirrhosis.

### Effects of comorbidities on Child Pugh Score depending on the liver disease

In the first plot, we tried to see if there is a correlation between the Child Pugh Score and the Charlson Comorbidity Index, which predicts the mortality of patients suffering from comorbidities:

```{r}
Charlson_vs_CPS_LV <- ggplot(
  data = dat_aug_03,
  mapping = aes(x = CPS,
                y = Charlson_CI,
                color = Liver_disease,
                na.rm = TRUE)
) +
  geom_point(alpha = 0.3) +
  geom_smooth(method = "lm") +
  scale_color_viridis_d() +
  labs(
    title = "Relationship between Child Pugh Score and Charlson score",
    subtitle = "for each liver disease",
    x = "CPS",
    y = "Charlson CI"
  ) +
  facet_wrap(~Liver_disease) +
  theme(
    strip.background = element_blank(),
    strip.text.x = element_blank()
)

Charlson_vs_CPS_LV
```

On that plot, we can see that the relationship between CPS and Charlson CI can be totally different depending on the liver disease from which the patients suffer. There is no clear pattern that can be deducted from these graphs or from the plot of Charlson CI versus CPS. However, we can't conclude that they are absolutely not correlated.

### Effects of comorbidities on Child Pugh Score with all liver diseases included

```{r}
Charlson_vs_CPS <- ggplot(
  data = dat_aug_03,
  mapping = aes(x = CPS,
                y = Charlson_CI,
                na.rm = TRUE)
) +
  geom_point(alpha = 0.1) +
  scale_color_viridis_d() +
  geom_smooth(method = "lm")
  labs(
  title = "Relationship between Child Pugh Score and Charlson score",
  x = "CPS",
  y = "Charlson CI"
)

Charlson_vs_CPS
```

# Effects of different etiologies on Child Pugh Score

Because the comorbidities don't seem to be related to the CPS, we are now interested in investigating which etiologies are significantly more or less present in patients with different CPS. In order to do that, we use a linear regression model:

```{r}
df <- df_Etiology |>
  select(-c(Otherliverdiseaseetiology,
            PrimarySclerosingCholangitis))

df
```

### Creation of a first model using the 'lm()' function

```{r}
model <- lm(formula = Alcohol ~ cps,
   data = df)
model
```

calculation of the mean values of the presence of the etiology "Alcohol" stratified on Child Pugh Score:

```{r}
df |> 
  group_by(cps) |> 
  summarise(mean = mean(Alcohol))
```

### Summary of the model

```{r}
model |> summary()
```

By looking at the p-value, we can see if the variable "Alcohol" is statistically significant or not. p-value \> 0.05 so the variable is not statistically significant.

Pivoting the data to get a long dataset, so that all the etiologies are defined in one column:

```{r}
df_Etiology_long <- df |> 
  pivot_longer(
    cols = -c(ID, cps),
    names_to = "Etiologies",
    values_to = "Presence"
  )

df_Etiology_long
```

### Addition of a new variable model_object

This variable will be computed per etiology and will contain the result to the call to the 'map()' function.

```{r}
df_Etiology_long_nested <- df_Etiology_long |> 
  group_by(Etiologies) |> # group_by() function to work per etiology
  nest() |> 
  ungroup()

df_Etiology_long_nested |> 
  group_by(Etiologies) |> 
  mutate(
    model_object = map(.x = data,
                       .f = ~lm(formula = Presence ~ cps,
                            data = .x))
  ) |> 
  filter(Etiologies == "Alcohol") |>
  pull(model_object)
```

### Use of the broom-package to tidy the fitted model

```{r}
df_Etiology_long_nested <- df_Etiology_long_nested|> 
  group_by(Etiologies) |> 
  mutate(
    model_object = map(.x = data,
                       .f = ~lm(formula = Presence ~ cps,
                                data = .x))
  )

df_Etiology_long_nested |>
  pull(model_object) |>
  pluck(1) |>
  tidy(conf.int = TRUE,
       conf.level = 0.95)
```

To apply the 'tidy()' function per model_object:

```{r}
df_Etiology_long_nested <- df_Etiology_long_nested |> 
  mutate(model_object_tidy = map(.x = model_object,
                                 .f = ~tidy(conf.int = TRUE,
                                           conf.level = 0.95,
                                           x = .x)
  )
)
df_Etiology_long_nested
```

Use of the 'unnest()' function to unpack the model_object_tidy:

```{r}
estimate <- df_Etiology_long_nested |>
  unnest( model_object_tidy
  )

estimate
```

Un-grouping the data:

```{r}
estimate <- estimate|> 
  filter(term == "cps") |> 
  ungroup(Etiologies)

estimate
```

Adding the variable q-value and an indicator variable denoting if a given etiology is significant or not:

```{r}
estimate <- estimate |> 
  mutate(
    q.value = p.adjust(p.value),
    is.significant = case_when(
      p.value < 0.05 ~ "yes",
      p.value >= 0.05 ~ "no"
    )
  )

estimate
```

### Visualization

```{r}
if (!requireNamespace("forcats", quietly = TRUE)) {
  install.packages("forcats")
}
library(forcats)


boxplot_model_etiologies <-  ggplot( 
  data = estimate,
  mapping = aes(x = estimate,
                  y = Etiologies)) +
  geom_errorbarh(aes(xmin = conf.low,
                 xmax = conf.high,
                 y = fct_reorder(Etiologies,estimate))) +
  geom_point()+
  geom_hline(yintercept = 0, linetype = "dashed", color = "black", size = 0.5) +
  labs(
    title = "Etiologies associated with CPS ",
    x = "Estimates(95% CIs)",
    y ="") +
  theme(plot.title = element_text(size = 12.5))

boxplot_model_etiologies
```

```{r}
Significance_plot_CPS <- estimate |> 
  select(q.value <= 0.05) |> 
  mutate(label = Etiologies) |> 
  ggplot(data = estimate, 
         mapping = aes(x = estimate, 
                       y = - log10(p.value), 
                       color = is.significant, 
                       label = ifelse(- log10(p.value) > 1,
                                      Etiologies,
                                      ""))) +
  geom_point(alpha = 0.8) +
  geom_text(hjust=0.5, vjust=-0.5) +
  labs(
    title = "Etiologies associated with CPS ",
    x = "Estimates(95% CIs)",
    y ="")  +
  theme(plot.title = element_text(hjust = 0.5, face = "bold", size = 14)) +
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=14,face="bold"))

Significance_plot_CPS

```

### Save plot

```{r}
ggsave("Significance_plot_CPS.png", 
       path = "../results/plots/",
       plot = Significance_plot_CPS,
       width = 7,
       height = 5)
```

In these plots, we can see that only one etiology is considered as having a significant correlation with CPS: Non-Alcoholic Fatty Liver Disease (NAFLD). Therefore, a patient with this etiology has more chances of having a low CPS than a patient with any other etiology.
