==> R/08_analysis_liver_disease.qmd <==
---
title: "08_analysis_liver_disease"
format: html
editor: visual
---

# This model turned out to be a dead end, we misunderstood how the linear regression was working and we used 2 categorical values without numerical value.

```{r}
library("tidyverse")
library("ggpubr")
library("forcats")
library("purrr")
library("broom")
library("ggplot2")
library("dplyr")
```

```{r}
dat_aug_03 <- read_tsv("~/projects/Group_19_project/data/03_dat_aug.tsv")
```

# Impact of different liver diseases on mortality

### Data widening

```{r}
df <- dat_aug_03 |>
  mutate(`Mortality (1 = y/0 = n)` = ifelse(Mortality == "No death", 0, 1)) |> 
  pivot_wider(
    names_from = Liver_disease,
    values_from = Liver_disease
    ) |> 
  mutate(across(10:19, ~ifelse(is.na(.), 0, 1)))
df
```

### Creation of a first model using the 'lm()' function

```{r}
model <- lm(formula = Ascites ~ `Mortality (1 = y/0 = n)`,
   data = df)
model
```

calculation of the mean values of the presence of the liver disease "Ascites" stratified on Mortality:

```{r}
df |> 
  group_by(`Mortality (1 = y/0 = n)`) |> 
  summarise(mean = mean(Ascites))
```

### Summary of the model

```{r}
model |> summary()
```

by looking at the p-value, we can see if the variable "Ascites" is statistically significant or not. p-value \< 0.05 so the variable is statistically significant.

Pivoting the data to get a long dataset, so that all the liver diseases are defined in one column:

```{r}
df_long <- df |> 
  pivot_longer(
    cols = c(`Hepatic encephalopathy`, 
             SBP,
             Ascites,
             `Other fluid overload`,
             AKI,
             `UGIB due to portal hypertension`,
             `HCC or cholangioca.`,
             `Alcoholic hepatitis`,
             `Jaundice- as isolated presentation`,
             `Hepato-hydrothorax`
             ),
    names_to = "Liver diseases",
    values_to = "Presence"
  )

df_long
```

### Addition of a new variable model_object

This variable will be computed per Liver disease and will contain the result to the call to the 'map()' function.

```{r}
df_long_nested <- df_long |> 
  group_by(`Liver diseases`) |> 
  nest() |> 
  ungroup()

df_long_nested |> 
  group_by(`Liver diseases`) |> 
  mutate(
    model_object = map(.x = data,
                       .f = ~lm(formula = Presence ~ `Mortality (1 = y/0 = n)`,
                            data = .x))
  ) |> 
  filter(`Liver diseases` == "Ascites") |>
  pull(model_object)
```

### Use of the broom-package to tidy the fitted model

```{r}
df_long_nested <- df_long_nested|> 
  group_by(`Liver diseases`) |> 
  mutate(
    model_object = map(.x = data,
                       .f = ~lm(formula = Presence ~ `Mortality (1 = y/0 = n)`,
                                data = .x))
  )

df_long_nested |>
  pull(model_object) |>
  pluck(1) |>
  tidy(conf.int = TRUE,
       conf.level = 0.95)
```

To apply the 'tidy()' function per model_object:

```{r}
df_long_nested <- df_long_nested |> 
  mutate(model_object_tidy = map(.x = model_object,
                                 .f = ~tidy(conf.int = TRUE,
                                           conf.level = 0.95,
                                           x = .x)
  )
)
df_long_nested
```

Use of the 'unnest()' function to unpack the model_object_tidy:

```{r}
estimate <- df_long_nested |>
  unnest( model_object_tidy)

estimate
```

Un-grouping the data:

```{r}
estimate <- estimate|> 
  filter(term == "`Mortality (1 = y/0 = n)`") |> 
  ungroup(`Liver diseases`)

estimate
```

Adding the variable q-value and an indicator variable denoting if a given liver disease is significant or not:

```{r}
estimate <- estimate |> 
  mutate(
    q.value = p.adjust(p.value),
    is.significant = case_when(
      p.value < 0.05 ~ "yes",
      p.value >= 0.05 ~ "no"
    )
  )

estimate
```

### Visualization

```{r}
if (!requireNamespace("forcats", quietly = TRUE)) {
  install.packages("forcats")
}
library(forcats)

estimate

Estimates_LV_Mortality <- estimate |> 
  ggplot( 
    mapping = aes(x = estimate,
                  y = `Liver diseases`)) +
  geom_errorbarh(aes(xmin = conf.low,
                 xmax = conf.high,
                 y = fct_reorder(`Liver diseases`,estimate))) +
  geom_point()+
  geom_hline(yintercept = 0, linetype = "dashed", color = "black", size = 0.5) +
  labs(
    title = "Liver diseases associated with mortality ",
    x = "Estimates(95% CIs)",
    y ="") +
  theme(plot.title = element_text(size = 12.5))
```

```{r}
Significance_plot_LV_mortality <- estimate |> 
  select(q.value <= 0.05) |> 
  mutate(
    label = `Liver diseases`
    ) |> 
  ggplot(
    data = estimate,
    mapping = aes(x = estimate, 
                  y = - log10(p.value), 
                  color = is.significant, 
                  label = ifelse(- log10(p.value) > 1,
                                 `Liver diseases`,
                                 "")
                  ) 
    ) +
  geom_point(alpha = 0.8) +
  geom_text(hjust=0.5, vjust=-0.5) +
  labs(
    title = "Liver diseases associated with mortality",
    x = "Estimates(95% CIs)",
    y ="") +
  scale_fill_viridis_d() +
  expand_limits(x = c(-0.05,0.12))

Significance_plot_LV_mortality
```

According to these two plots, three of the observed liver diseases have are significantly associated with mortality. These diseases are Hepato-hydrothorax, hepatic encephalopathy and jaundice as isolated presentation. They all have a positive association with mortality, which means that the patients with these liver diseases are more likely to die of cirrhosis than the patients with other liver diseases. These results are coherent with the plot of the "liver disease repartition in cyrrhosis patients, stratified by mortality" present in the file '05_analysis\_ mortality.qmd'.
